fix(uma/mace): audit fixes for calculators, runners, engine & docs
Comprehensive audit of the UMA and MACE modules fixing theoretical
methodology and computational-principle bugs, plus documentation sync.

Critical fixes:
- MACE head kwarg: use singular `head` (str), not `heads` (list); the
  MACECalculator API reads `kwargs.get("head")` (mace_calc.py).
- Adsorption energy: derive gas_task automatically from PBC so the
  isolated gas molecule is scored with a molecular task (UMA periodic ->
  omol; generic engine -> molecule). `gas_task` param overrides (api.py).
- MD temperature: delegate to ASE atoms.get_temperature(), which is
  self-consistent with force_temperature/thermostat and the constraint
  removed_dof accounting (no FixCom double-count). FixSymmetry falls back
  to 3N-3 with a warning (md.py).
- NVE pre_relax: default off (energy-conserving, pre-relaxing moves to a
  0K minimum and changes the conserved-energy baseline); NVT stays on.
  Removed pre_relax from BUILTIN_DEFAULTS so the resolver no longer forces
  it True on the NVE production path; engine owns the ensemble-aware
  default (engine.py, defaults.py).

Robustness:
- NaN/inf abort via _check_finite() wired into SP/Opt/MD runners.
- Optimization re-raises "Non-finite" RuntimeErrors instead of swallowing.
- MACE dtype is calc-type aware: float32 for MD, float64 for sp/opt.
- Stress computation gated on atoms.pbc.any() (skip for molecules).
- Molecular spin/charge defaults are engine-aware: omol gets charge=0/
  spin=1 (multiplicity); generic molecule only defaults charge=0 (no spin,
  avoiding MACE total_spin mismatch).
- Vacuum cell for cell-less molecules so CONTCAR/OUTCAR writers don't crash.

Documentation (CN/EN synced, CLI help, INCAR template):
- Correct molecule charge/spin semantics (omol multiplicity vs MACE total
  spin S); omol charge/spin is "default 0/1" not "required".
- Pre-relaxation defaults: NVT on / NVE off, with physical rationale.
- Document MACE --head / --dtype, gas_task auto-derivation, NaN safety,
  and molecular-stress gating.
- Add PRE_RELAX / DEFAULT_DTYPE / HEAD to INCAR reference tables.

Tests: 21 regression tests in tests/core/mlipx/test_fixes.py covering all
fixes; full suite 136 passed, 1 skipped.

+++ uma/mlipx/calculators/factory.py
@@ -151,14 +151,15 @@ class CalculatorFactory:
         elif m_type == "mace":
             from mlipx.calculators.mace_calc import MACECalculatorWrapper  # noqa: PLC0415
 
-            # Plan section 11.1 / 12: factory-level dtype fallback is float32
-            # (recommended for long MD on V100). The wrapper constructor still
-            # defaults to float64 for direct/test construction.
+            # Dtype fallback for *direct* factory construction is float64 (the
+            # safe high-precision default). The engine overrides this with a
+            # calc-type-aware default (float32 for MD, float64 for sp/opt) when
+            # nothing in the config layer already set default_dtype.
             return MACECalculatorWrapper(
                 model_path=model_path,
                 device=device,
                 task=task,
-                default_dtype=kwargs.get("default_dtype", "float32"),
+                default_dtype=kwargs.get("default_dtype", "float64"),
                 head=kwargs.get("head"),
             )
         elif m_type == "dpa":
